+++ rana_qgis_plugin/widgets/rana_browser.py
@@ -5,7 +5,7 @@ from enum import Enum
 from pathlib import Path
 from typing import List
 
-from qgis.PyQt.QtCore import QModelIndex, QSettings, Qt, pyqtSignal, pyqtSlot
+from qgis.PyQt.QtCore import QModelIndex, QSettings, Qt, QTimer, pyqtSignal, pyqtSlot
 from qgis.PyQt.QtGui import QAction, QPixmap, QStandardItem, QStandardItemModel
 from qgis.PyQt.QtSvg import QSvgWidget
 from qgis.PyQt.QtWidgets import (
@@ -57,6 +57,7 @@ from rana_qgis_plugin.utils_api import (
 class RevisionsView(QWidget):
     new_simulation_clicked = pyqtSignal(int)
     create_3di_model_clicked = pyqtSignal(int)
+    open_schematisation_revision_in_qgis_requested = pyqtSignal(dict, dict)
     busy = pyqtSignal()
     ready = pyqtSignal()
 
@@ -69,10 +70,6 @@ class RevisionsView(QWidget):
         self.setup_ui()
 
     def setup_ui(self):
-        revisions_refreash_btn = QToolButton()
-        revisions_refreash_btn.setToolTip("Refresh")
-        revisions_refreash_btn.clicked.connect(self.show_revisions)
-        revisions_refreash_btn.setIcon(refresh_icon)
         self.revisions_table = QTableView()
         self.revisions_table.setSortingEnabled(True)
         self.revisions_table.verticalHeader().hide()
@@ -83,8 +80,11 @@ class RevisionsView(QWidget):
         self.revisions_table.horizontalHeader().setSectionResizeMode(
             QHeaderView.Stretch
         )
+        self.revisions_table.setContextMenuPolicy(
+            Qt.ContextMenuPolicy.CustomContextMenu
+        )
+        self.revisions_table.customContextMenuRequested.connect(self.menu_requested)
         layout = QVBoxLayout(self)
-        layout.addWidget(revisions_refreash_btn)
         layout.addWidget(self.revisions_table)
         self.setLayout(layout)
 
@@ -93,11 +93,31 @@ class RevisionsView(QWidget):
         self.selected_file = selected_file
         self.show_revisions()
 
+    def menu_requested(self, pos):
+        index = self.revisions_table.indexAt(pos)
+        revision_item = self.revisions_model.itemFromIndex(index)
+        if not revision_item:
+            return
+
+        threedi_revision, schematisation = revision_item.data()
+        if threedi_revision:
+            menu = QMenu(self)
+            action = QAction("Open in QGIS", self)
+            action.triggered.connect(
+                lambda _: self.open_schematisation_revision_in_qgis_requested.emit(
+                    threedi_revision.to_dict(), schematisation["schematisation"]
+                )
+            )
+            menu.addAction(action)
+        menu.popup(self.revisions_table.viewport().mapToGlobal(pos))
+
+    def refresh(self):
+        self.show_revisions()
+
     def show_revisions(self):
         self.busy.emit()
         selected_file = self.selected_file
-        self.revisions_model.clear()
-        # collect rows to show in widget, format: [date_str, event, (button_label, signal_func)]
+        # collect rows to show in widget, format: [date_str, event, (button_label, signal_func), revision, schematisation]
         rows = []
         if selected_file.get("data_type") == "threedi_schematisation":
             # retrieve schematisation and revisions
@@ -129,21 +149,44 @@ class RevisionsView(QWidget):
                         "Create 3Di model",
                         lambda _: self.create_3di_model_clicked.emit(revision.id),
                     )
-                rows.append([date_str, revision.commit_message, btn_data])
+                rows.append(
+                    [
+                        date_str,
+                        revision.commit_message,
+                        btn_data,
+                        revision,
+                        schematisation,
+                    ]
+                )
         else:
             history = get_tenant_project_file_history(
                 self.project["id"], {"path": self.selected_file["id"]}
             )
             for item in history["items"]:
                 date_str = convert_to_local_time(item["created_at"])
-                rows.append([date_str, item["message"], None])
+                rows.append([date_str, item["message"], None, None, None])
 
         # Populate table
+        sort_column = self.revisions_table.horizontalHeader().sortIndicatorSection()
+        sort_order = self.revisions_table.horizontalHeader().sortIndicatorOrder()
+        self.revisions_model.clear()
         self.revisions_model.setHorizontalHeaderLabels(["Timestamp", "Event", ""])
-        for i, (date_str, event, btn_data) in enumerate(rows):
+        for i, (
+            date_str,
+            event,
+            btn_data,
+            threedi_revision,
+            threedi_schematisation,
+        ) in enumerate(rows):
+            date_item = QStandardItem(date_str)
+
+            # We store the revision object for loading specific revisions in menu_requested.
+            if threedi_revision:
+                date_item.setData((threedi_revision, threedi_schematisation))
+
             self.revisions_model.appendRow(
                 [
-                    QStandardItem(date_str),
+                    date_item,
                     QStandardItem(event),
                     QStandardItem(""),
                 ]
@@ -155,6 +198,8 @@ class RevisionsView(QWidget):
                 self.revisions_table.setIndexWidget(
                     self.revisions_model.index(i, 2), btn
                 )
+        self.revisions_table.sortByColumn(sort_column, sort_order)
+        self.revisions_table.setSortingEnabled(True)
         self.ready.emit()
 
 
@@ -173,10 +218,6 @@ class FileView(QWidget):
         self.project = project
 
     def setup_ui(self):
-        file_refresh_btn = QToolButton()
-        file_refresh_btn.setToolTip("Refresh")
-        file_refresh_btn.clicked.connect(self.refresh)
-        file_refresh_btn.setIcon(refresh_icon)
         self.file_table_widget = QTableWidget(1, 2)
         self.file_table_widget.horizontalHeader().setVisible(False)
         self.file_table_widget.verticalHeader().setVisible(False)
@@ -191,14 +232,12 @@ class FileView(QWidget):
         button_layout.addWidget(self.btn_create_model)
         button_layout.addWidget(self.btn_show_revisions)
         layout = QVBoxLayout(self)
-        layout.addWidget(file_refresh_btn)
         layout.addWidget(self.file_table_widget)
         layout.addLayout(button_layout)
         self.setLayout(layout)
 
     def show_selected_file_details(self, selected_file):
         self.selected_file = selected_file
-        self.file_table_widget.clearContents()
         schematisation_button = None
         self.btn_create_model.hide()
         self.btn_start_simulation.hide()
@@ -274,6 +313,7 @@ class FileView(QWidget):
                 file_details.extend(schematisation_details)
             else:
                 self.communication.show_error("Failed to download 3Di schematisation.")
+        self.file_table_widget.clearContents()
         self.file_table_widget.setRowCount(len(file_details))
         self.file_table_widget.horizontalHeader().setStretchLastSection(True)
         for i, (label, value) in enumerate(file_details):
@@ -333,10 +373,6 @@ class FilesBrowser(QWidget):
         self.fetch_and_populate(project)
 
     def setup_ui(self):
-        project_refresh_btn = QToolButton()
-        project_refresh_btn.setToolTip("Refresh")
-        project_refresh_btn.setIcon(refresh_icon)
-        project_refresh_btn.clicked.connect(self.update)
         self.files_tv = QTreeView()
         self.files_tv.setContextMenuPolicy(Qt.ContextMenuPolicy.CustomContextMenu)
         self.files_tv.customContextMenuRequested.connect(self.menu_requested)
@@ -347,14 +383,14 @@ class FilesBrowser(QWidget):
         self.files_tv.doubleClicked.connect(self.select_file_or_directory)
         self.btn_upload = QPushButton("Upload Files to Rana")
         layout = QVBoxLayout(self)
-        # todo: move refresh to far right (some day)
-        layout.addWidget(project_refresh_btn)
         layout.addWidget(self.files_tv)
         layout.addWidget(self.btn_upload)
         self.setLayout(layout)
 
     def refresh(self):
-        self.update()
+        if self.selected_item["type"] == "directory":
+            self.fetch_and_populate(self.project, self.selected_item["id"])
+        self.communication.clear_message_bar()
 
     def update(self):
         selected_path = self.selected_item["id"]
@@ -421,10 +457,11 @@ class FilesBrowser(QWidget):
             project["id"],
             {"path": path} if path else None,
         )
+        sort_column = self.files_tv.header().sortIndicatorSection()
+        sort_order = self.files_tv.header().sortIndicatorOrder()
         self.files_model.clear()
         header = ["Filename", "Data type", "Size", "Last modified"]
         self.files_model.setHorizontalHeaderLabels(header)
-
         directories = [file for file in self.files if file["type"] == "directory"]
         files = [file for file in self.files if file["type"] == "file"]
 
@@ -467,6 +504,9 @@ class FilesBrowser(QWidget):
                 [name_item, data_type_item, size_item, last_modified_item]
             )
 
+        self.files_tv.sortByColumn(sort_column, sort_order)
+        self.files_tv.setSortingEnabled(True)
+
         for i in range(len(header)):
             self.files_tv.resizeColumnToContents(i)
         self.files_tv.setColumnWidth(0, 300)
@@ -501,11 +541,6 @@ class ProjectsBrowser(QWidget):
         self.projects_search = QLineEdit()
         self.projects_search.setPlaceholderText("🔍 Search for project by name")
         self.projects_search.textChanged.connect(self.filter_projects)
-        # Create refresh button
-        overview_refresh_btn = QToolButton()
-        overview_refresh_btn.setToolTip("Refresh")
-        overview_refresh_btn.clicked.connect(self.refresh)
-        overview_refresh_btn.setIcon(refresh_icon)
         # Create tree view with project files and model
         self.projects_model = QStandardItemModel()
         self.projects_tv = QTreeView()
@@ -524,7 +559,6 @@ class ProjectsBrowser(QWidget):
         # Organize widgets in layouts
         top_layout = QHBoxLayout()
         top_layout.addWidget(self.projects_search)
-        top_layout.addWidget(overview_refresh_btn)
         pagination_layout = QHBoxLayout()
         pagination_layout.addWidget(self.btn_previous)
         pagination_layout.addWidget(self.label_page_number)
@@ -768,12 +802,16 @@ class RanaBrowser(QWidget):
     start_simulation_selected_with_revision = pyqtSignal(dict, dict, int)
     create_model_selected = pyqtSignal(dict)
     create_model_selected_with_revision = pyqtSignal(dict, int)
+    open_schematisation_selected_with_revision = pyqtSignal(dict, dict)
     delete_file_selected = pyqtSignal(dict, dict)
 
     def __init__(self, communication: UICommunication):
         super().__init__()
         self.communication = communication
         self.setup_ui()
+        self.refresh_timer = QTimer()
+        self.refresh_timer.timeout.connect(self.auto_refresh)
+        self.refresh_timer.start(60000)
 
     @property
     def project(self):
@@ -792,13 +830,19 @@ class RanaBrowser(QWidget):
         self.rana_processes = QWidget()
         self.rana_files = QStackedWidget()
         self.rana_browser.addTab(self.rana_files, "Files")
-        self.rana_browser.addTab(self.rana_processes, "Processes")
+        # self.rana_browser.addTab(self.rana_processes, "Processes")
         self.rana_browser.setCurrentIndex(0)
-        self.rana_browser.setTabEnabled(1, False)
+        # self.rana_browser.setTabEnabled(1, False)
         # Set up breadcrumbs, browser and file view widgets
         self.breadcrumbs = BreadCrumbsWidget(
             communication=self.communication, parent=self
         )
+        refresh_btn = QToolButton()
+        refresh_btn.setToolTip("Refresh")
+        refresh_btn.setIcon(refresh_icon)
+        refresh_btn.clicked.connect(self.refresh)
+        self.rana_browser.setCornerWidget(refresh_btn, Qt.TopRightCorner)
+
         # Setup top layout with logo and breadcrumbs
         top_layout = QHBoxLayout()
 
@@ -915,6 +959,11 @@ class RanaBrowser(QWidget):
                 self.project, self.selected_item, revision_id
             )
         )
+        # Load specific revision of schematisation
+        self.revisions_view.open_schematisation_revision_in_qgis_requested.connect(
+            self.open_schematisation_selected_with_revision
+        )
+
         # Ensure correct page is shown - do this last zo all updates are done
         self.projects_browser.projects_refreshed.connect(
             lambda: self.rana_files.setCurrentIndex(0)
@@ -949,6 +998,11 @@ class RanaBrowser(QWidget):
     @pyqtSlot()
     def disable(self):
         self.rana_browser.setEnabled(False)
+
+    def auto_refresh(self):
+        # skip auto refresh for projects view to not mess up pagination
+        if self.rana_files.currentIndex() in [1, 2, 3]:
+            self.refresh()
 
     @pyqtSlot()
     def refresh(self):