+++ HISTORY.rst
@@ -8,6 +8,8 @@ History
 - Replace file interaction buttons with context menu
 - Download generated postprocessing rasters.
 - Add button to generate threedimodel when missing
+- Replace page specific refresh buttons with single refresh button
+- Refresh files browser, file view and revisions view every 60 seconds
 
 
 1.1.20 (2025-11-18)